remove EVMClient, ContractDeployer, etc
