+++ integration-tests/actions/actions.go
@@ -4,9 +4,7 @@ package actions
 import (
 	"context"
 	"crypto/ecdsa"
-	"encoding/hex"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"math/big"
 	"os"
@@ -22,135 +20,16 @@ import (
 	"github.com/go-resty/resty/v2"
 	"github.com/rs/zerolog"
 
-	"github.com/ethereum/go-ethereum"
-	"github.com/ethereum/go-ethereum/accounts/keystore"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/google/uuid"
-	"github.com/rs/zerolog/log"
-	"go.uber.org/zap/zapcore"
-
 	"github.com/smartcontractkit/seth"
 
-	"github.com/smartcontractkit/chainlink-testing-framework/blockchain"
 	ctfClient "github.com/smartcontractkit/chainlink-testing-framework/client"
-	"github.com/smartcontractkit/chainlink-testing-framework/k8s/environment"
-	"github.com/smartcontractkit/chainlink-testing-framework/logging"
-	"github.com/smartcontractkit/chainlink-testing-framework/testreporters"
-	"github.com/smartcontractkit/chainlink-testing-framework/utils/conversions"
 	"github.com/smartcontractkit/chainlink-testing-framework/utils/testcontext"
 	"github.com/smartcontractkit/chainlink/integration-tests/client"
 	"github.com/smartcontractkit/chainlink/integration-tests/contracts"
 	"github.com/smartcontractkit/chainlink/integration-tests/docker/test_env"
 )
-
-// ContractDeploymentInterval After how many contract actions to wait before starting any more
-// Example: When deploying 1000 contracts, stop every ContractDeploymentInterval have been deployed to wait before continuing
-var ContractDeploymentInterval = 200
-
-// FundChainlinkNodes will fund all of the provided Chainlink nodes with a set amountCreateOCRv2Jobs of native currency
-// Deprecated: we are moving away from blockchain.EVMClient, use actions_seth.FundChainlinkNodes
-func FundChainlinkNodes(
-	nodes []*client.ChainlinkK8sClient,
-	client blockchain.EVMClient,
-	amount *big.Float,
-) error {
-	for _, cl := range nodes {
-		toAddress, err := cl.PrimaryEthAddress()
-		if err != nil {
-			return err
-		}
-		recipient := common.HexToAddress(toAddress)
-		msg := ethereum.CallMsg{
-			From:  common.HexToAddress(client.GetDefaultWallet().Address()),
-			To:    &recipient,
-			Value: conversions.EtherToWei(amount),
-		}
-		gasEstimates, err := client.EstimateGas(msg)
-		if err != nil {
-			return err
-		}
-		err = client.Fund(toAddress, amount, gasEstimates)
-		if err != nil {
-			return err
-		}
-	}
-	return client.WaitForEvents()
-}
-
-// FundChainlinkNodesAddress will fund all of the provided Chainlink nodes address at given index with a set amount of native currency
-func FundChainlinkNodesAddress(
-	nodes []*client.ChainlinkK8sClient,
-	client blockchain.EVMClient,
-	amount *big.Float,
-	keyIndex int,
-) error {
-	for _, cl := range nodes {
-		toAddress, err := cl.EthAddresses()
-		if err != nil {
-			return err
-		}
-		toAddr := common.HexToAddress(toAddress[keyIndex])
-		gasEstimates, err := client.EstimateGas(ethereum.CallMsg{
-			To: &toAddr,
-		})
-		if err != nil {
-			return err
-		}
-		err = client.Fund(toAddress[keyIndex], amount, gasEstimates)
-		if err != nil {
-			return err
-		}
-	}
-	return client.WaitForEvents()
-}
-
-// FundChainlinkNodesAddress will fund all of the provided Chainlink nodes addresses with a set amount of native currency
-func FundChainlinkNodesAddresses(
-	nodes []*client.ChainlinkClient,
-	client blockchain.EVMClient,
-	amount *big.Float,
-) error {
-	for _, cl := range nodes {
-		toAddress, err := cl.EthAddressesForChain(client.GetChainID().String())
-		if err != nil {
-			return err
-		}
-		for _, addr := range toAddress {
-			toAddr := common.HexToAddress(addr)
-			gasEstimates, err := client.EstimateGas(ethereum.CallMsg{
-				To: &toAddr,
-			})
-			if err != nil {
-				return err
-			}
-			err = client.Fund(addr, amount, gasEstimates)
-			if err != nil {
-				return err
-			}
-		}
-	}
-	return client.WaitForEvents()
-}
-
-// FundChainlinkNodes will fund all of the provided Chainlink nodes with a set amount of native currency
-func FundChainlinkNodesLink(
-	nodes []*client.ChainlinkK8sClient,
-	blockchain blockchain.EVMClient,
-	linkToken contracts.LinkToken,
-	linkAmount *big.Int,
-) error {
-	for _, cl := range nodes {
-		toAddress, err := cl.PrimaryEthAddress()
-		if err != nil {
-			return err
-		}
-		err = linkToken.Transfer(toAddress, linkAmount)
-		if err != nil {
-			return err
-		}
-	}
-	return blockchain.WaitForEvents()
-}
 
 // ChainlinkNodeAddresses will return all the on-chain wallet addresses for a set of Chainlink nodes
 func ChainlinkNodeAddresses(nodes []*client.ChainlinkK8sClient) ([]common.Address, error) {
@@ -264,242 +143,11 @@ func GetMockserverInitializerDataForOTPE(
 	return initializers, nil
 }
 
-// TeardownSuite tears down networks/clients and environment and creates a logs folder for failed tests in the
-// specified path. Can also accept a testreporter (if one was used) to log further results
-func TeardownSuite(
-	t *testing.T,
-	env *environment.Environment,
-	chainlinkNodes []*client.ChainlinkK8sClient,
-	optionalTestReporter testreporters.TestReporter, // Optionally pass in a test reporter to log further metrics
-	failingLogLevel zapcore.Level, // Examines logs after the test, and fails the test if any Chainlink logs are found at or above provided level
-	grafnaUrlProvider testreporters.GrafanaURLProvider,
-	clients ...blockchain.EVMClient,
-) error {
-	l := logging.GetTestLogger(t)
-	if err := testreporters.WriteTeardownLogs(t, env, optionalTestReporter, failingLogLevel, grafnaUrlProvider); err != nil {
-		return fmt.Errorf("Error dumping environment logs, leaving environment running for manual retrieval, err: %w", err)
-	}
-	// Delete all jobs to stop depleting the funds
-	err := DeleteAllJobs(chainlinkNodes)
-	if err != nil {
-		l.Warn().Msgf("Error deleting jobs %+v", err)
-	}
-
-	for _, c := range clients {
-		if c != nil && chainlinkNodes != nil && len(chainlinkNodes) > 0 {
-			if err := ReturnFunds(chainlinkNodes, c); err != nil {
-				// This printed line is required for tests that use real funds to propagate the failure
-				// out to the system running the test. Do not remove
-				fmt.Println(environment.FAILED_FUND_RETURN)
-				l.Error().Err(err).Str("Namespace", env.Cfg.Namespace).
-					Msg("Error attempting to return funds from chainlink nodes to network's default wallet. " +
-						"Environment is left running so you can try manually!")
-			}
-		} else {
-			l.Info().Msg("Successfully returned funds from chainlink nodes to default network wallets")
-		}
-		// nolint
-		if c != nil {
-			err := c.Close()
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return env.Shutdown()
-}
-
-// TeardownRemoteSuite is used when running a test within a remote-test-runner, like for long-running performance and
-// soak tests
-// Deprecated: we are moving away from blockchain.EVMClient, use actions_seth.TeardownRemoteSuite
-func TeardownRemoteSuite(
-	t *testing.T,
-	namespace string,
-	chainlinkNodes []*client.ChainlinkK8sClient,
-	optionalTestReporter testreporters.TestReporter, // Optionally pass in a test reporter to log further metrics
-	grafnaUrlProvider testreporters.GrafanaURLProvider,
-	client blockchain.EVMClient,
-) error {
-	l := logging.GetTestLogger(t)
-	var err error
-	if err = testreporters.SendReport(t, namespace, "./", optionalTestReporter, grafnaUrlProvider); err != nil {
-		l.Warn().Err(err).Msg("Error writing test report")
-	}
-	// Delete all jobs to stop depleting the funds
-	err = DeleteAllJobs(chainlinkNodes)
-	if err != nil {
-		l.Warn().Msgf("Error deleting jobs %+v", err)
-	}
-
-	if err = ReturnFunds(chainlinkNodes, client); err != nil {
-		l.Error().Err(err).Str("Namespace", namespace).
-			Msg("Error attempting to return funds from chainlink nodes to network's default wallet. " +
-				"Environment is left running so you can try manually!")
-	}
-	return err
-}
-
-func DeleteAllJobs(chainlinkNodes []*client.ChainlinkK8sClient) error {
-	for _, node := range chainlinkNodes {
-		if node == nil {
-			return fmt.Errorf("found a nil chainlink node in the list of chainlink nodes while tearing down: %v", chainlinkNodes)
-		}
-		jobs, _, err := node.ReadJobs()
-		if err != nil {
-			return fmt.Errorf("error reading jobs from chainlink node, err: %w", err)
-		}
-		for _, maps := range jobs.Data {
-			if _, ok := maps["id"]; !ok {
-				return fmt.Errorf("error reading job id from chainlink node's jobs %+v", jobs.Data)
-			}
-			id := maps["id"].(string)
-			_, err := node.DeleteJob(id)
-			if err != nil {
-				return fmt.Errorf("error deleting job from chainlink node, err: %w", err)
-			}
-		}
-	}
-	return nil
-}
-
-// ReturnFunds attempts to return all the funds from the chainlink nodes and other wallets to the network's default wallet,
-// which will always be the first wallet in the list of wallets. If errors are encountered, it will keep trying other wallets
-// and return all errors encountered.
-func ReturnFunds(chainlinkNodes []*client.ChainlinkK8sClient, blockchainClient blockchain.EVMClient) error {
-	if blockchainClient == nil {
-		return fmt.Errorf("blockchain client is nil, unable to return funds from chainlink nodes")
-	}
-	log.Info().Msg("Attempting to return Chainlink node funds to default network wallets")
-	if blockchainClient.NetworkSimulated() {
-		log.Info().Str("Network Name", blockchainClient.GetNetworkName()).
-			Msg("Network is a simulated network. Skipping fund return.")
-		return nil
-	}
-
-	// If we fail to return funds from some addresses, we still want to try to return funds from the rest
-	encounteredErrors := []error{}
-
-	if len(blockchainClient.GetWallets()) > 1 {
-		if err := blockchainClient.SetDefaultWallet(0); err != nil {
-			encounteredErrors = append(encounteredErrors, err)
-		} else {
-			for walletIndex := 1; walletIndex < len(blockchainClient.GetWallets()); walletIndex++ {
-				decodedKey, err := hex.DecodeString(blockchainClient.GetWallets()[walletIndex].PrivateKey())
-				if err != nil {
-					encounteredErrors = append(encounteredErrors, err)
-					continue
-				}
-				privKey, err := crypto.ToECDSA(decodedKey)
-				if err != nil {
-					encounteredErrors = append(encounteredErrors, err)
-					continue
-				}
-
-				err = blockchainClient.ReturnFunds(privKey)
-				if err != nil {
-					encounteredErrors = append(encounteredErrors, err)
-					continue
-				}
-			}
-		}
-	}
-
-	for _, chainlinkNode := range chainlinkNodes {
-		fundedKeys, err := chainlinkNode.ExportEVMKeysForChain(blockchainClient.GetChainID().String())
-		if err != nil {
-			encounteredErrors = append(encounteredErrors, err)
-			continue
-		}
-		for _, key := range fundedKeys {
-			keyToDecrypt, err := json.Marshal(key)
-			if err != nil {
-				encounteredErrors = append(encounteredErrors, err)
-				continue
-			}
-			// This can take up a good bit of RAM and time. When running on the remote-test-runner, this can lead to OOM
-			// issues. So we avoid running in parallel; slower, but safer.
-			decryptedKey, err := keystore.DecryptKey(keyToDecrypt, client.ChainlinkKeyPassword)
-			if err != nil {
-				encounteredErrors = append(encounteredErrors, err)
-				continue
-			}
-			err = blockchainClient.ReturnFunds(decryptedKey.PrivateKey)
-			if err != nil {
-				encounteredErrors = append(encounteredErrors, fmt.Errorf("error returning funds from chainlink node: %w", err))
-				continue
-			}
-		}
-	}
-	if err := blockchainClient.WaitForEvents(); err != nil {
-		encounteredErrors = append(encounteredErrors, err)
-	}
-	if len(encounteredErrors) > 0 {
-		return fmt.Errorf("encountered errors while returning funds: %v", encounteredErrors)
-	}
-	return nil
-}
-
-// FundAddresses will fund a list of addresses with an amount of native currency
-func FundAddresses(blockchain blockchain.EVMClient, amount *big.Float, addresses ...string) error {
-	for _, address := range addresses {
-		toAddr := common.HexToAddress(address)
-		gasEstimates, err := blockchain.EstimateGas(ethereum.CallMsg{
-			To: &toAddr,
-		})
-		if err != nil {
-			return err
-		}
-		if err := blockchain.Fund(address, amount, gasEstimates); err != nil {
-			return err
-		}
-	}
-	return blockchain.WaitForEvents()
-}
-
 // EncodeOnChainExternalJobID encodes external job uuid to on-chain representation
 func EncodeOnChainExternalJobID(jobID uuid.UUID) [32]byte {
 	var ji [32]byte
 	copy(ji[:], strings.Replace(jobID.String(), "-", "", 4))
 	return ji
-}
-
-// UpgradeChainlinkNodeVersions upgrades all Chainlink nodes to a new version, and then runs the test environment
-// to apply the upgrades
-func UpgradeChainlinkNodeVersions(
-	testEnvironment *environment.Environment,
-	newImage, newVersion string,
-	nodes ...*client.ChainlinkK8sClient,
-) error {
-	if newImage == "" || newVersion == "" {
-		return errors.New("New image and new version is needed to upgrade the node")
-	}
-	for _, node := range nodes {
-		if err := node.UpgradeVersion(testEnvironment, newImage, newVersion); err != nil {
-			return err
-		}
-	}
-	err := testEnvironment.RunUpdated(len(nodes))
-	if err != nil { // Run the new environment and wait for changes to show
-		return err
-	}
-	return client.ReconnectChainlinkNodes(testEnvironment, nodes)
-}
-
-func DeployLINKToken(cd contracts.ContractDeployer) (contracts.LinkToken, error) {
-	linkToken, err := cd.DeployLinkTokenContract()
-	if err != nil {
-		return nil, err
-	}
-	return linkToken, err
-}
-
-func DeployMockETHLinkFeed(cd contracts.ContractDeployer, answer *big.Int) (contracts.MockETHLINKFeed, error) {
-	mockETHLINKFeed, err := cd.DeployMockETHLINKFeed(answer)
-	if err != nil {
-		return nil, err
-	}
-	return mockETHLINKFeed, err
 }
 
 // todo - move to CTF
